requirements_lite.txt
# Minimal requirements for fast Colab training

# Core
torch>=2.0.0
torchvision>=0.15.0

# Data
Pillow>=9.0.0
numpy>=1.24.0

# Training
tqdm>=4.65.0

# Optional (for Colab)
kagglehub  # For dataset download
